+++ src/app/layout.tsx
@@ -2,6 +2,7 @@ import type { Metadata } from "next";
 import { Inter } from "next/font/google";
 import "./globals.css";
 import Footer from "@/components/home/Footer/Footer";
+import ScrollButton from "@/components/home/ScrollButton/ScrollButton";
 
 const inter = Inter({ subsets: ["latin"] });
 
@@ -20,6 +21,7 @@ export default function RootLayout({
       <body className={inter.className}>
         {children}
         <Footer />
+        <ScrollButton />
       </body>
     </html>
   );
