+++ src/components/home/Footer/Footer.tsx
@@ -1,7 +1,5 @@
 "use client";
-import { Box, Button, Container, Flex, Text } from "@chakra-ui/react";
 import React from "react";
-import { Icon } from "@chakra-ui/react";
 import { ImFacebook } from "react-icons/im";
 import { strings } from "@/assets/locales/locales";
 import FooterGridItem from "./FooterGridItem";
@@ -25,38 +23,28 @@ const Footer = () => {
               duration: 0.5,
               delay: 0.2,
             }}
-            // fontSize="35px"
-            // textAlign="center"
-            // whiteSpace="pre-wrap"
             className="text-[35px] text-center whitespace-pre-wrap"
           >
-            {/* Simplify management, save time, and grow your
-            <br /> business. */}
             {footerData?.footerHeader}
           </motion.h1>
 
           <div className="flex flex-col md:flex-row items-center justify-center md:space-x-4 lg:space-x-6">
             <button className="bg-tertiary hover:bg-[#97CE2C] text-secondary px-10 py-4 my-5 rounded">
-              {/* Start Your Free Trial */}
               {footerData?.footerBtn1}
             </button>
             <button className="bg-[#2574a3] hover:bg-[#4C7DB6] text-secondary px-10 py-4 my-5 rounded">
-              {/* Take a Quick Tour */}
               {footerData?.footerBtn2}
             </button>
           </div>
         </div>
-        <hr className="w-full my-7"></hr>
+        <hr className="w-full my-7" />
         <div className="grid grid-cols-1 md:grid-cols-5 gap-2 pl-2">
           {footerColumns?.map((column, index) => (
             <FooterGridItem key={index} column={column} />
           ))}
         </div>
         <div className="flex justify-between py-7 flex-col md:flex-row">
-          <h1>
-            {/* © Teachworks 2021 */}
-            {footerData?.footerCopyRight}
-          </h1>
+          <h1>{footerData?.footerCopyRight}</h1>
           <div className="mr-[200px]">
             <button className="rounded-full text-secondary bg-transparent hover:bg-secondary hover:text-black p-3">
               <a
