+++ src/components/home/Footer/FooterGridItem.tsx
@@ -5,14 +5,8 @@ const FooterGridItem = ({ column }: { column: any }) => {
     <React.Fragment>
       <div>
         <h2 className="font-semibold mb-3">{column.footerTitle}</h2>
-
         {column.footerLinks.map((link: string) => (
           <p
-            // cursor="pointer"
-            // mb="10px"
-            // key={link}
-            // textColor="footerLinksColor"
-            // _hover={{ textColor: "secondary" }}
             className="cursor-pointer mb-3 text-footerLinksColor hover:text-secondary font-extralight"
             key={link}
           >
